additional fixes

+++ pages/api/availability/eventtype.ts
@@ -144,11 +144,13 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         const openingHours = req.body.availability.openingHours || [];
         // const overrides = req.body.availability.dateOverrides || [];
 
-        await prisma.availability.deleteMany({
-          where: {
-            eventTypeId: +req.body.id,
-          },
-        });
+        if (+req.body.id) {
+          await prisma.availability.deleteMany({
+            where: {
+              eventTypeId: +req.body.id,
+            },
+          });
+        }
 
         Promise.all(
           openingHours.map((schedule) =>
@@ -178,12 +180,6 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
   if (req.method == "DELETE") {
     await prisma.eventTypeCustomInput.deleteMany({
-      where: {
-        eventTypeId: req.body.id,
-      },
-    });
-
-    await prisma.webhookEventTypes.deleteMany({
       where: {
         eventTypeId: req.body.id,
       },
